% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsuperClassAnalysis.R
\docType{package}
\name{unsuperClassAnalysis}
\alias{unsuperClassAnalysis}
\alias{unsuperClassAnalysis-package}
\title{Toolbox for supervised classifications}
\description{
This R package is designed to maximize the classification accuracy for a single class using an
unsupervised classification approach. The functions are based on the
\code{\link[RStoolbox]{unsuperClass}} function from the
\href{https://bleutner.github.io/RStoolbox/}{RStoolbox} package by
Benjamin Leutner \url{benjamin.leutner@uni-wuerzburg.de}.
}
\details{
Function included in this package:
\itemize{
  \item \code{\link{unsupSaturation}}
}
}
\references{
This program was written at Remote Sensing Solutions RSS GmbH,
            Dingolfinger Str. 9, 81673 Munich, Germany \url{https://rssgmbh.de/}.
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
